juanbarearojo#9 Cambia el tipo de 'cantidad' a uint64 para reducir pruebas
- Actualiza el campo 'cantidad' de float32 a uint64, eliminando la necesidad de verificar que 'cantidad' sea mayor que 0 en el constructor.
- Al ser un entero sin signo, se garantiza que siempre sea no negativo, lo que simplifica la lógica del código.
- Reduce la carga de trabajo del programador al evitar pruebas adicionales para validar valores negativos.

closes juanbarearojo#9

+++ internal/models/ingrediente.go
@@ -7,7 +7,7 @@ import (
 
 type Ingrediente struct {
 	productos      map[string]Producto
-	cantidad       float32
+	cantidad       uint64
 	fechaCaducidad *string
 }
